requirements.txt
# This is the minimal set of requirements to run all the apis,
# Some of the agents, esp. GPT or BERT require more dependencies. Install them from agents/requirements.txt

# Tested on Python 3.6 or 3.7.
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
darkskylib==0.3.91
eventlet==0.17.4
Flask==0.12.2
Flask-Cors==3.0.8
Flask-SocketIO==4.2.1
frozendict==1.2
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-speech==1.3.0
google-cloud-texttospeech==0.5.0
google-cloud-translate==2.0.0
googleapis-common-protos==1.6.0
googlemaps==3.1.4
grpcio==1.25.0
humanfriendly==4.18
idna==2.9
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
nltk==3.4.5
prompt-toolkit==3.0.4
protobuf==3.11.1
pyasn1-modules==0.2.7
python-engineio==3.10.0
python-socketio==4.3.1
pytz==2019.3
rapidfuzz==0.10.0
requests==2.23.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==0.16.0
wit==5.1.0
